+++ acceptance/acceptance_test.go
@@ -74,6 +74,11 @@ const (
 	MaxFileSize      = 100_000
 	// Filename to save replacements to (used by diff.py)
 	ReplsFile = "repls.json"
+
+	// ENVFILTER allows filtering subtests matching certain env var.
+	// e.g. ENVFILTER=SERVERLESS=yes will run all tests that run SERVERLESS to "yes"
+	// The tests the don't set SERVERLESS variable or set to empty string will also be run.
+	EnvFilterVar = "ENVFILTER"
 )
 
 var Scripts = map[string]bool{
@@ -213,6 +218,8 @@ func testAccept(t *testing.T, inprocessMode bool, singleTest string) int {
 	totalDirs := 0
 	selectedDirs := 0
 
+	envFilters := getEnvFilters(t)
+
 	for _, dir := range testDirs {
 		totalDirs += 1
 
@@ -244,7 +251,7 @@ func testAccept(t *testing.T, inprocessMode bool, singleTest string) int {
 				if len(expanded[0]) > 0 {
 					t.Logf("Running test with env %v", expanded[0])
 				}
-				runTest(t, dir, 0, coverDir, repls.Clone(), config, configPath, expanded[0], inprocessMode)
+				runTest(t, dir, 0, coverDir, repls.Clone(), config, configPath, expanded[0], inprocessMode, envFilters)
 			} else {
 				for ind, envset := range expanded {
 					envname := strings.Join(envset, "/")
@@ -252,7 +259,7 @@ func testAccept(t *testing.T, inprocessMode bool, singleTest string) int {
 						if !inprocessMode {
 							t.Parallel()
 						}
-						runTest(t, dir, ind, coverDir, repls.Clone(), config, configPath, envset, inprocessMode)
+						runTest(t, dir, ind, coverDir, repls.Clone(), config, configPath, envset, inprocessMode, envFilters)
 					})
 				}
 			}
@@ -262,6 +269,27 @@ func testAccept(t *testing.T, inprocessMode bool, singleTest string) int {
 	t.Logf("Summary (dirs): %d/%d/%d run/selected/total, %d skipped", selectedDirs-skippedDirs, selectedDirs, totalDirs, skippedDirs)
 
 	return len(testDirs)
+}
+
+func getEnvFilters(t *testing.T) []string {
+	envFilterValue := os.Getenv(EnvFilterVar)
+	if envFilterValue == "" {
+		return nil
+	}
+
+	filters := strings.Split(envFilterValue, ",")
+
+	for _, filter := range filters {
+		items := strings.Split(filter, "=")
+		if len(items) != 2 || len(items[0]) == 0 {
+			t.Fatalf("Invalid filter %q in %s=%q", filter, EnvFilterVar, envFilterValue)
+		}
+		key := items[0]
+		// Clear it just to be sure, since it's going to be part of os.Environ() and we're going to add different value based on settings.
+		os.Unsetenv(key)
+	}
+
+	return filters
 }
 
 func getTests(t *testing.T) []string {
@@ -353,6 +381,7 @@ func runTest(t *testing.T,
 	configPath string,
 	customEnv []string,
 	inprocessMode bool,
+	envFilters []string,
 ) {
 	if LogConfig {
 		configBytes, err := json.MarshalIndent(config, "", "  ")
@@ -439,7 +468,7 @@ func runTest(t *testing.T,
 	// Must be added PrepareReplacementsUser, otherwise conflicts with [USERNAME]
 	testdiff.PrepareReplacementsUUID(t, &repls)
 
-	// User replacements come last:
+	// User replacements:
 	repls.Repls = append(repls.Repls, config.Repls...)
 
 	// Save replacements to temp test directory so that it can be read by diff.py
@@ -475,7 +504,24 @@ func runTest(t *testing.T,
 		require.Len(t, items, 2)
 		key := items[0]
 		value := items[1]
-		cmd.Env = addEnvVar(t, cmd.Env, &repls, key, value, config.EnvRepl, len(config.EnvMatrix[key]) > 1)
+		// Only add replacement by default if value is part of EnvMatrix with more than 1 option and length is 4 or more chars
+		// (to avoid matching "yes" and "no" values from template input parameters)
+		cmd.Env = addEnvVar(t, cmd.Env, &repls, key, value, config.EnvRepl, len(config.EnvMatrix[key]) > 1 && len(value) >= 4)
+	}
+
+	for filterInd, filterEnv := range envFilters {
+		filterEnvKey := strings.Split(filterEnv, "=")[0]
+		for ind := range cmd.Env {
+			// Search backwards, because the latest settings is what is actually applicable.
+			envPair := cmd.Env[len(cmd.Env)-1-ind]
+			if strings.Split(envPair, "=")[0] == filterEnvKey {
+				if envPair == filterEnv {
+					break
+				} else {
+					t.Skipf("Skipping because test environment %s does not match ENVFILTER#%d: %s", envPair, filterInd, filterEnv)
+				}
+			}
+		}
 	}
 
 	absDir, err := filepath.Abs(dir)
@@ -955,7 +1001,7 @@ func getNodeTypeID(cloudEnv string) string {
 	case "gcp":
 		return "n1-standard-4"
 	case "":
-		return "local-fake-node"
+		return "i3.xlarge"
 	default:
 		return "nodetype-" + cloudEnv
 	}