+++ bundle/config/validate/validate_sync_patterns.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"fmt"
 	"strings"
-	"sync"
 
 	"github.com/databricks/cli/bundle"
 	"github.com/databricks/cli/libs/diag"
@@ -43,9 +42,8 @@ func (v *validateSyncPatterns) Apply(ctx context.Context, b *bundle.Bundle) diag
 }
 
 func checkPatterns(patterns []string, path string, b *bundle.Bundle) (diag.Diagnostics, error) {
-	var mu sync.Mutex
 	var errs errgroup.Group
-	var diags diag.Diagnostics
+	var diags diag.SafeDiagnostics
 
 	for index, pattern := range patterns {
 		// If the pattern is negated, strip the negation prefix
@@ -67,18 +65,16 @@ func checkPatterns(patterns []string, path string, b *bundle.Bundle) (diag.Diagn
 
 			if len(all) == 0 {
 				path := fmt.Sprintf("%s[%d]", path, index)
-				mu.Lock()
-				diags = diags.Append(diag.Diagnostic{
+				diags.Append(diag.Diagnostic{
 					Severity:  diag.Warning,
 					Summary:   fmt.Sprintf("Pattern %s does not match any files", pattern),
 					Locations: b.Config.GetLocations(path),
 					Paths:     []dyn.Path{dyn.MustPathFromString(path)},
 				})
-				mu.Unlock()
 			}
 			return nil
 		})
 	}
 
-	return diags, errs.Wait()
+	return diags.Diags, errs.Wait()
 }