+++ libs/structdiff/bench.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+set -ex
+exec go test -bench=. -benchmem -run ^x