+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Version changelog
 
+## Release v0.254.0
+
+### Bundles
+* Added `experimental.skip_artifact_cleanup` flag ([#2980](https://github.com/databricks/cli/pull/2980))
+* Add an experimental project template for Lakeflow Declarative Pipelines ([#2959](https://github.com/databricks/cli/pull/2959))
+
+
+## Release v0.253.0
+
+### Dependency updates
+* Upgrade SDK to v0.70.0 ([#2920](https://github.com/databricks/cli/pull/2920))
+* Upgrade TF provider to v1.81.0 ([#2936](https://github.com/databricks/cli/pull/2936))
+
+
 ## Release v0.252.0
 
 ### Dependency updates